[ADD] 프레임1,2 컴포넌트 분리작업
