+++ src/pages/Main/Frame1.scss
@@ -1,0 +1,108 @@
+.frame_1 {
+  margin-top: 1rem;
+  max-width: 55%;
+  width: 55%;
+  height: 80%;
+}
+
+.calendar {
+  margin: 0 2rem;
+  padding: 2rem 2rem 1rem 2rem;
+  border: 0.5px solid #626060;
+  border-radius: 15px;
+
+  &-label {
+    margin: 0 0 1rem 2rem;
+    font-weight: 600;
+    font-size: 20px;
+    color: #c62917;
+  }
+
+  &-table {
+    width: 100%;
+    height: 100%;
+    text-align: center;
+    border-spacing: 0px;
+    margin: auto;
+  }
+
+  &-day {
+    text-align: center;
+    background-color: #eeeeee;
+    width: 7rem;
+    height: 3rem;
+    border: 1px solid #e0e0e0;
+    border-radius: 0px;
+
+    &.right_up_corner {
+      border-radius: 0px 15px 0px 0px;
+    }
+  }
+
+  &-class {
+    width: 7rem;
+    height: 4rem;
+    border: 1px solid #e0e0e0;
+    border-radius: 0px;
+
+    &.right_down_corner {
+      border-radius: 0px 0px 15px 0px;
+    }
+  }
+
+  &-number {
+    text-align: center;
+    width: 2rem;
+    height: 3rem;
+    border: 1px solid #e0e0e0;
+    border-radius: 0px;
+    &.left_up_corner {
+      height: 1.5rem;
+      border-radius: 15px 0px 0px 0px;
+    }
+    &.left_down_corner {
+      border-radius: 0px 0px 0px 15px;
+    }
+  }
+
+  &-info {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem 0 0 0;
+
+    &-calculator {
+      max-width: 12rem;
+      width: 40%;
+      border: 0.5px solid #e0e0e0;
+      border-radius: 15px;
+      font-weight: 400;
+      color: #626060;
+      font-size: 14px;
+      padding: 0.3rem 1rem;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-around;
+    }
+
+    &-buttons {
+      height: 100%;
+      width: 60%;
+      display: flex;
+      justify-content: space-around;
+      align-items: center;
+
+      &-button {
+        font-size: 18px;
+        margin-left: 1rem;
+        color: #ffffff;
+        max-width: 6rem;
+        width: 40%;
+        height: 30%;
+        background: #a31432;
+        border: none;
+        border-radius: 15px;
+      }
+    }
+  }
+}
