+++ src/pages/Main/Frame2.scss
@@ -1,0 +1,64 @@
+.frame_2 {
+    margin-top: 1rem;
+    max-width: 45%;
+    width: 45%;
+    height: 80%;
+  }
+
+  .sort_selecter {
+    margin: 0 2rem 1rem 2rem;
+    text-align: end;
+    &-select {
+      text-align: center;
+      margin: 0;
+      width: 8rem;
+      height: 1.5rem;
+      border: 1px solid #626060;
+      border-radius: 15px;
+      background-color: #ffffff;
+      &-img {
+        height: 1.2rem;
+      }
+    }
+  }
+  
+  .filter {
+    margin: 0 2rem 1rem 2rem;
+    text-align: end;
+    :hover {
+      background-color: #a31432;
+    }
+    &-button {
+      margin: 0;
+      width: 5.5rem;
+      height: max-content;
+      border: 1px solid #626060;
+      border-radius: 15px;
+      background-color: #ffffff;
+      &-img {
+        height: 1.2rem;
+      }
+    }
+  }
+  
+  .sort_selecter {
+    margin: 1rem 2rem;
+    text-align: end;
+  }
+  
+  .lecture_list {
+    padding-bottom: 1rem;
+    :hover {
+      background-color: #a31432;
+      color: #ffffff;
+    }
+  }
+  
+  .content2 {
+    width: 100%;
+    height: 100%;
+    border: 0.5px solid #626060;
+    border-radius: 15px;
+    overflow: scroll;
+  }
+
